Deserialize and serialize objects internally in UbloxGps instead of in callbacks.

+++ app/src/main/java/ca/indigogames/ubloxagps/ublox/messages/UbxClassId.java
@@ -1,13 +1,15 @@
 package ca.indigogames.ubloxagps.ublox.messages;
 
+import ca.indigogames.ubloxagps.compat.UByte;
+
 public class UbxClassId {
-    public static final byte NAV = 0x01;
-    public static final byte RXM = 0x02;
-    public static final byte INF = 0x04;
-    public static final byte ACK = 0x05;
-    public static final byte CFG = 0x06;
-    public static final byte MON = 0x0A;
-    public static final byte AID = 0x0B;
-    public static final byte TIM = 0x0D;
-    public static final byte LOG = 0x21;
+    public static final UByte NAV = UByte.valueOf((byte)0x01);
+    public static final UByte RXM = UByte.valueOf((byte)0x02);
+    public static final UByte INF = UByte.valueOf((byte)0x04);
+    public static final UByte ACK = UByte.valueOf((byte)0x05);
+    public static final UByte CFG = UByte.valueOf((byte)0x06);
+    public static final UByte MON = UByte.valueOf((byte)0x0A);
+    public static final UByte AID = UByte.valueOf((byte)0x0B);
+    public static final UByte TIM = UByte.valueOf((byte)0x0D);
+    public static final UByte LOG = UByte.valueOf((byte)0x21);
 }
